Moved Main Character and Implemented Perspective
This push moves the MainCharacter into PlaygroundCharacter (as it is simpler for current needs), and MainCharacter into TestCharacter to better signify it is for testing rather than usage.

Second implementation is PerspectiveManager, which is the current solution for dynamic perspectives for an actor and their movement. It will eventually be replaced by Delegate Callbacks when the crashing in UE is fixed, but for now Blueprints crash when attempt event bindings in Blueprints to a c++ function.