Procedural Foliage Work
Started an implementation of procedural foliage. i.e. dynamic animations handled in the shaders rather than in a mesh animation.  This implementation is based off UE4, which has had some changes since UE5 that makes it more clunky. It is working however, but will be changed.

Possible candidates will be the GoT implementation (Ghost of Tsushima), which will require extra work with a compute shader.

The other is one based off of  Vertex Coloring, which will augment the situation we have currently. It probably will generalize better to things like swaying trees and more complex meshes, but that's to be seen.