Updated PlaygroundCharacter with States
This adds the states used in our test character to Playground character in preparation for new animation blueprints. Current goal is to implement the proper animations between Idle, Walking, and Running.

May need to create a custom CharacterMovementComponent for good design on managing running v.s. walking. However, it may not really need to be necessary, and may not be worth it. Still deciding on the best practice at the moment.