Multiple additions
Spellcasting system has been updated; It has a base for the general setup for how all spells will be cast in the future. Animations are still in the works, just haven't had time to make them better.

More additions to the village area; The first section has been roughly figured out with some details to make the upcoming enemy placements be more interesting.

Inventory and Items implemented. Items can now be picked up and placed in the world. Only visual is a particle effect at the moment. Looking into the glow method from souls games to add a stylistic approach.