Pause menu and Spell menu implemented
The base for the pause menu and inventory have been made. Currently, you can now pick up spells from the environment and it'll be added to the menu dynamically.

I may get to this in a bit here, but the next steps are to make it so clicking on the spell in the inventory will added it to the equipped spells list in the spellbook (this is a debug feature only currently, equipped spells will eventually come with the souls style method). Once there, I will then change the debug setup of casting spells to then use the spellbook instead of directly using the water orb.