Implemented Attack System Base
Began the implementation of the attack system based off of the casting system.

This system has some issues with some timings for some animations, and may run into issues later on when the animations are improved in quality.

However, for now the system is currently starting an attack, activating the weapon as necessary and overlapping.

Next steps to do:
- Improve animation quality
- implement enemy UI for damage (going to go with a dark souls style health bar)
- Improve the RPG elements to account for damage types.
- Implement damage to enemies (damage to player happens, but enemies don't currently have it).