Reorganized Player State Machine code
Cleaned up the player State Machine code to use fewer parameters and to access parent variables for necessary data. This cleans up the general code and makes it look much nicer.

Also settled on the general state machine design for attacking, guarding, and deflection, which will be implemented next. The current design is to have Attacking, guarding, and deflection be the primary states for combat states. Guarding willl allow walking, like casting. Deflection is a dual state, which can refer to both deflecting and being deflected, controlled by variables. Attacking is as its named, and all three states utilize a CanChain  variable as a way of staying within those states, and chaining attacks and deflects in an organized manner without the need to juggle between Idle states and attacking states.