Deflection System Started
The Deflection System has been put in with the attack system, and offensive deflections can occur to the character (i.e. the character can hit a deflection box, and be deflected).

Currently, there are no full recovery animations, so chaining doesn't work fully, and that's the next step for the whole system. Chaining will require more animations (and some improved ones), as well as full recovery animations for deflections so the character doesn't go into an Idle state right after deflecting.