Added Combo Attacks to our character.
Attack combos have been added. So far there is only a 2-chain combo available with basic animations.

There is possibly a bug for attacks not correctly ending animation-wise; It's happened a few times, but is not consistently recreatable. It's possible it was fixed, but something worth noting for later to work on if it happens again. Generally it happens when spamming the attack button, but I think I've seen it when not pressing anything, so difficult to say.